Added custom HttpClient class MyHttpClient, that support self-signed ssl certificates for HTTPS.
Instead of using:

	new DefaultHttpClient()

for setting up an HttpClient, use:

	new MyHttpClient(null)

everything else is the same.

ERRORS:

There is a problem with registering a new user. After entering and submitting the form, an error gets thrown on the call

	String successMsg = JSONresponse.get("messages").toString();

in RegisterActivity.java. It throws "JSONException: no value for messages". I've marked the line with a comment, line 114.